Separate out lock server into separate file
