+++ dsync-server_test.go
@@ -1,0 +1,120 @@
+/*
+ * Minio Cloud Storage, (C) 2016 Minio, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package dsync_test
+
+import (
+	"errors"
+	"fmt"
+	. "github.com/minio/dsync"
+	"sync"
+	"time"
+)
+
+// used when cached timestamp do not match with what client remembers.
+var errInvalidTimestamp = errors.New("Timestamps don't match, server may have restarted.")
+
+const WriteLock = -1
+
+type lockServer struct {
+	mutex sync.Mutex
+	// Map of locks, with negative value indicating (exclusive) write lock
+	// and positive values indicating number of read locks
+	lockMap   map[string]int64
+	timestamp time.Time // Timestamp set at the time of initialization. Resets naturally on minio server restart.
+}
+
+func (l *lockServer) verifyArgs(args *LockArgs) error {
+	if !l.timestamp.Equal(args.Timestamp) {
+		return errInvalidTimestamp
+	}
+	// disabled for test framework
+	//if !isRPCTokenValid(args.Token) {
+	//	return errInvalidToken
+	//}
+	return nil
+}
+
+func (l *lockServer) Lock(args *LockArgs, reply *bool) error {
+	l.mutex.Lock()
+	defer l.mutex.Unlock()
+	if err := l.verifyArgs(args); err != nil {
+		return err
+	}
+	if _, *reply = l.lockMap[args.Name]; !*reply {
+		l.lockMap[args.Name] = WriteLock // No locks held on the given name, so claim write lock
+	}
+	*reply = !*reply // Negate *reply to return true when lock is granted or false otherwise
+	return nil
+}
+
+func (l *lockServer) Unlock(args *LockArgs, reply *bool) error {
+	l.mutex.Lock()
+	defer l.mutex.Unlock()
+	if err := l.verifyArgs(args); err != nil {
+		return err
+	}
+	var locksHeld int64
+	if locksHeld, *reply = l.lockMap[args.Name]; !*reply { // No lock is held on the given name
+		return fmt.Errorf("Unlock attempted on an unlocked entity: %s", args.Name)
+	}
+	if *reply = locksHeld == WriteLock; !*reply { // Unless it is a write lock
+		return fmt.Errorf("Unlock attempted on a read locked entity: %s (%d read locks active)", args.Name, locksHeld)
+	}
+	delete(l.lockMap, args.Name) // Remove the write lock
+	return nil
+}
+
+const ReadLock = 1
+
+func (l *lockServer) RLock(args *LockArgs, reply *bool) error {
+	l.mutex.Lock()
+	defer l.mutex.Unlock()
+	if err := l.verifyArgs(args); err != nil {
+		return err
+	}
+	var locksHeld int64
+	if locksHeld, *reply = l.lockMap[args.Name]; !*reply {
+		l.lockMap[args.Name] = ReadLock // No locks held on the given name, so claim (first) read lock
+		*reply = true
+	} else {
+		if *reply = locksHeld != WriteLock; *reply { // Unless there is a write lock
+			l.lockMap[args.Name] = locksHeld + ReadLock // Grant another read lock
+		}
+	}
+	return nil
+}
+
+func (l *lockServer) RUnlock(args *LockArgs, reply *bool) error {
+	l.mutex.Lock()
+	defer l.mutex.Unlock()
+	if err := l.verifyArgs(args); err != nil {
+		return err
+	}
+	var locksHeld int64
+	if locksHeld, *reply = l.lockMap[args.Name]; !*reply { // No lock is held on the given name
+		return fmt.Errorf("RUnlock attempted on an unlocked entity: %s", args.Name)
+	}
+	if *reply = locksHeld != WriteLock; !*reply { // A write-lock is held, cannot release a read lock
+		return fmt.Errorf("RUnlock attempted on a write locked entity: %s", args.Name)
+	}
+	if locksHeld > ReadLock {
+		l.lockMap[args.Name] = locksHeld - ReadLock // Remove one of the read locks held
+	} else {
+		delete(l.lockMap, args.Name) // Remove the (last) read lock
+	}
+	return nil
+}
